requirements.txt
agentscope==0.0.6a2
aiofiles==23.2.1
aiohttp==3.8.6
aiolimiter==1.1.0
aioredis==2.0.1
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
anthropic==0.18.1
anyio==4.3.0
anytree==2.12.1
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bce-python-sdk==0.9.6
beautifulsoup4==4.12.3
bidict==0.23.1
black==24.4.2
blinker==1.7.0
cachetools==5.3.3
camel-converter==3.1.1
certifi==2024.2.2
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
configparser==7.0.0
dashscope==1.14.1
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
Django==4.2.11
dnspython==2.6.1
docstring_parser==0.16
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
Farama-Notifications==0.0.4
fastjsonschema==2.19.1
filelock==3.13.4
fire==0.4.0
flake8==7.1.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gitdb==4.0.11
gitignore_parser==0.1.9
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
grpcio-tools==1.62.1
gymnasium==0.29.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
hyperframe==6.0.1
idna==3.7
ijson==3.3.0
imap-tools==1.5.0
importlib_metadata==7.1.0
inputimeout==1.0.4
ipykernel==6.27.1
ipython==8.17.2
ipywidgets==8.1.1
iso8601==1.1.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-path==0.3.2
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
lancedb==0.4.0
lazy-object-proxy==1.10.0
libcst==1.0.1
litellm==1.40.25
llama-index==0.10.29
llama-index-agent-openai==0.2.2
llama-index-cli==0.1.11
llama-index-core==0.10.29
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.15
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.19
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.1
llamaindex-py-client==0.1.18
loguru==0.6.0
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
meilisearch==0.21.0
more-itertools==10.2.0
motor==3.4.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.9.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.24.3
ollama==0.1.8
openai==1.35.3
openapi-core==0.18.2
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
openpyxl==3.1.2
overrides==7.7.0
packaging==24.0
pandas==2.1.1
parse==1.20.1
parso==0.8.4
pathable==0.4.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
playwright==1.43.0
portalocker==2.8.2
prance==23.6.21.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycryptodome==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
pyee==11.1.0
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pylance==0.9.0
pymongo==4.6.3
pypdf==4.2.0
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.0
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qdrant-client==1.7.0
qianfan==0.3.2
rank-bm25==0.2.2
ratelimiter==1.2.0.post0
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
retry==0.9.2
rfc3339-validator==0.1.4
rich==13.6.0
rpds-py==0.18.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.3.2
scipy==1.13.0
semantic-kernel==0.4.3.dev0
semver==3.0.2
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.4.4
stack-data==0.6.3
striprtf==0.0.26
ta==0.10.2
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.4.0
tiktoken==0.7.0
tokenizers==0.15.2
tomli==2.0.1
tornado==6.4
tortoise-orm==0.21.3
tqdm==4.66.2
traitlets==5.14.2
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.18
wcwidth==0.2.13
websocket-client==1.6.2
websockets==11.0.3
Werkzeug==3.0.2
widgetsnbextension==4.0.10
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.4
zhipuai==2.0.1
zipp==3.18.1
